Cleanup poorly designed hacks.
I remember being totally swamped and in a hurry when writing this one. Now (years later) I did the "what was I thinking" thing and cleaned it up. Moved the dimming code to the base displayObj where it really should be. Dumped the wierd begin thing back to normal begin() method.